+++ src/Twig/NodeVisitor/ModelNodeVisitor.php
@@ -46,10 +46,19 @@ final class ModelNodeVisitor implements FindPresentationModelInterface, NodeVisi
      */
     public function leaveNode(Node $node, Environment $environment): Node
     {
-        if ($node instanceof ModuleNode) {
-            $node->setNode('display_start', new DisplayStartNode($node->getTemplateName(), $this));
-            $node->setNode('display_end', new DisplayEndNode($node->getTemplateName()));
+        if (!($node instanceof ModuleNode)) {
+            return $node;
         }
+
+        if ($node->hasAttribute('embedded_templates')) {
+            /** @var ModuleNode $embeddedTemplate */
+            foreach ($node->getAttribute('embedded_templates') as $embeddedTemplate) {
+                $embeddedTemplate->setAttribute('is_embedded', true);
+            }
+        }
+
+        $node->setNode('display_start', new DisplayStartNode($node, $this));
+        $node->setNode('display_end', new DisplayEndNode($node));
 
         return $node;
     }